refactor: hard cutover — remove backcompat re-exports
Removed re-exports from:
- typescript/provider-sdk/src/protocol.ts (ProtocolType, ProtocolTypeValue, ProtocolSmallestUnit)
- typescript/infra/src/govern/HyperlaneAppGovernor.ts (AnnotatedCallData, InferredCall, SubmissionType)
- typescript/infra/src/utils/safe.ts (getSafeApiKey)

Callers now import directly from the canonical module. Added provider-sdk to the changeset.
