+++ typescript/infra/scripts/safes/governance/update-signers.ts
@@ -9,8 +9,8 @@ import {
   getGovernanceSafes,
   getGovernanceSigners,
 } from '../../../config/environments/mainnet3/governance/utils.js';
-import { AnnotatedCallData } from '../../../src/govern/HyperlaneAppGovernor.js';
 import { SafeMultiSend } from '../../../src/govern/multisend.js';
+import { AnnotatedCallData } from '../../../src/govern/types.js';
 import { withGovernanceType } from '../../../src/governance.js';
 import { GovernanceType } from '../../../src/governanceTypes.js';
 import { Role } from '../../../src/roles.js';
