+++ typescript/infra/src/govern/HyperlaneAppGovernor.ts
@@ -43,9 +43,6 @@ import {
 import type { AnnotatedCallData, InferredCall } from './types.js';
 import { SubmissionType } from './types.js';
 
-export type { AnnotatedCallData, InferredCall };
-export { SubmissionType };
-
 export abstract class HyperlaneAppGovernor<
   App extends HyperlaneApp<any>,
   Config extends OwnableConfig,
